+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support for App CRs with a `v` prefixed version. This enabled Flux to automatically update the version based on its image tag.
+- Support for App CRs with a `v` prefixed version. This enables Flux to automatically update the version based on its image tag.
 
 ## [6.0.0] - 2021-11-29