+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.2.0] - 2021-12-21
+
+### Added
+
+- Add `ValidateAppUpdate` to ensure `.spec.Namespace` is immutable in App CRs.
+
 ## [6.1.0] - 2021-12-09
 
 ### Added
@@ -281,7 +287,8 @@ since there can be a delay creating them on cluster creation.
 
 - First release
 
-[Unreleased]: https://github.com/giantswarm/app/compare/v6.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app/compare/v6.2.0...HEAD
+[6.2.0]: https://github.com/giantswarm/app/compare/v6.1.0...v6.2.0
 [6.1.0]: https://github.com/giantswarm/app/compare/v6.0.0...v6.1.0
 [6.0.0]: https://github.com/giantswarm/app/compare/v5.6.0...v6.0.0
 [5.6.0]: https://github.com/giantswarm/app/compare/v5.5.0...v5.6.0